Break out first few methods into modules for pool in dbus_api
