+++ src/dbus_api/pool/fetch_properties_2_0/api.rs
@@ -1,0 +1,31 @@
+use dbus::{
+    tree::{Factory, Method, MTFn},
+};
+
+use crate::dbus_api::{
+    pool::fetch_properties_2_0::methods::{get_properties, get_all_properties},
+    types::TData,
+};
+
+pub fn get_all_properties_method(f: &Factory<MTFn<TData>, TData>)
+        -> Method<MTFn<TData>, TData> {
+    f.method("GetAllProperties", (), get_all_properties)
+        // a{s(bv)}: Dictionary of property names to tuples
+        // In the tuple:
+        // b: Indicates whether the property value fetched was successful
+        // v: If b is true, represents the value for the given property
+        //    If b is false, represents the error returned when fetching the property
+        .out_arg(("results", "a{s(bv)}"))
+}
+
+pub fn get_properties_method(f: &Factory<MTFn<TData>, TData>)
+        -> Method<MTFn<TData>, TData> {
+    f.method("GetProperties", (), get_properties)
+        .in_arg(("properties", "as"))
+        // a{s(bv)}: Dictionary of property names to tuples
+        // In the tuple:
+        // b: Indicates whether the property value fetched was successful
+        // v: If b is true, represents the value for the given property
+        //    If b is false, represents the error returned when fetching the property
+        .out_arg(("results", "a{s(bv)}"))
+}
